Avoid calling .format() twice for log messages
Code in this plugin was using str.format() to create log messages, then
passing them to beets.log() which would call str.format() again. This
leads to crashes if our message contains a string like '{}' or '%s'
because the second call to str.format() will interpret them as control
codes.